requirements.txt
alembic==1.18.4
bidict==0.23.1
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
et_xmlfile==2.0.0
eventlet==0.40.4
Flask==3.1.2
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.6.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
openpyxl==3.1.5
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
python-engineio==4.13.0
python-slugify==8.0.4
python-socketio==5.16.0
requests==2.32.5
simple-websocket==1.1.0
SQLAlchemy==2.0.45
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wsproto==1.3.2
